backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
psycopg2-binary
mysqlclient 
pydantic[email]
"pydantic<2.0.0"
python-multipart